Refactor: Move network package inside vm_supervisor
